% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoUnits.R
\name{autoUnits}
\alias{autoUnits}
\title{Auto Time Interval Units}
\usage{
autoUnits(x)
}
\arguments{
\item{x}{an object inheriting from class \code{\link{difftime}}}
}
\description{
Convert time interval units to natural values based on 
  magnitude of difference.
}
\examples{
autoUnits(as.difftime("0:3:35"))
autoUnits(as.difftime("15:3:35"))
autoUnits(ISOdate(2000, 5, 1) - ISOdate(2000, 4, 20))

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
