% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing.R
\name{bearing}
\alias{bearing}
\title{Calculate Bearing Between Two Positions}
\usage{
bearing(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1, lon1}{numeric. The latitude and longitude of the starting coordinate in decimal degrees.}

\item{lat2, lon2}{numeric. The latitude and longitude of the ending coordinate in decimal degrees.}
}
\value{
vector with initial and final bearings.
}
\description{
Calculates the bearing between two points, given each point's latitude and longitude coordinates
}
\examples{
# What is the bearing from San Diego, CA to Honolulu, HI?
bearing(32.87, -117.25, 21.35, -157.98)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
