% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaParams.R
\name{betaParams}
\alias{betaParams}
\title{Calculate Beta parameters}
\usage{
betaParams(w, k)
}
\arguments{
\item{w}{mode}

\item{k}{concentration}
}
\description{
Calculate Beta shape parameters and variance from 
  mode and concentration.
}
