% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catSpatInterp.R
\name{catSpatInterp}
\alias{catSpatInterp}
\title{Categorical Spatial Interpolation}
\usage{
catSpatInterp(
  data,
  x.col = "x",
  y.col = "y",
  group.col = "group",
  num.grid = 100,
  knn = 10,
  hull.buffer = 0.1,
  num.cores = 1,
  num.batches = NULL
)
}
\arguments{
\item{data}{matrix or data.frame containing points and grouping designation.}

\item{x.col, y.col, group.col}{numbers or characters identifying which columns 
in \code{data} are the x and y values and grouping designation.}

\item{num.grid}{number of grid cells for k-nearest neighbor interpolation.}

\item{knn}{number of nearest neighbors to consider for interpolation.}

\item{hull.buffer}{percent increase of convex hull to use as spatial area to
interpolate over.}

\item{num.cores}{number of cores to distribute interpolations over.}

\item{num.batches}{number of batches to divide grid cell interpolations into.}
}
\value{
A list containing a raster and points of buffered convex hull.
}
\description{
Create a raster of probability of categorical values 
  interpolated across a 2-dimensional space given a set of points where 
  each is assigned to one of several classes.
}
\references{
Adapted from code originally presented in a blog post on 
  Categorical Spatial Interpolation by Timo Grossenbacher
  \url{https://timogrossenbacher.ch/2018/03/categorical-spatial-interpolation-with-r/}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
