% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central.quantile.R
\name{central.quantile}
\alias{central.quantile}
\title{Central Quantile}
\usage{
central.quantile(x, pct = 0.95)
}
\arguments{
\item{x}{numeric vector.}

\item{pct}{central percentile desired.}
}
\value{
a two element vector giving the lower and upper quantiles.
}
\description{
Upper and lower values of central quantile
}
\examples{
x <- runif(1000)
central.quantile(x)
central.quantile(x, pct = 0.75)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
