% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.name.R
\name{color.name}
\alias{color.name}
\title{Color Name}
\usage{
color.name(i)
}
\arguments{
\item{i}{integer specifying color .}
}
\value{
character value of 'i' color.
}
\description{
Return the name of a color listed given the number.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
