% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.distance.R
\name{convert.distance}
\alias{convert.distance}
\title{Distance Conversion}
\usage{
convert.distance(x, from = c("nm", "km", "mi"), to = c("km", "nm", "mi"))
}
\arguments{
\item{x}{numeric. The distance to be converted.}

\item{from, to}{character. Units to convert from and to. Can be "km" (kilometers), 
"nm" (nautical miles), or "mi" (statute miles), or any partial match 
thereof (case sensitive).}
}
\description{
Convert distances between kilometers, nautical miles, and statute miles.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
