% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.point.R
\name{crossing.point}
\alias{crossing.point}
\title{Crossing Point}
\usage{
crossing.point(l1, l2)
}
\arguments{
\item{l1, l2}{matrices representing two lines, where first two columns are
x and y values respectively}
}
\value{
a data.frame of x and y values of points where lines cross
}
\description{
Return point where two lines cross
}
\examples{
x <- 1:100
line1 <- cbind(x, 3 + 3 * x)
line2 <- cbind(x, 10 - 3 * x)
plot(line1[, 1], line1[, 2], type = "l", col = "red")
lines(line2[, 1], line2[, 2], col = "blue")
cr.pt <- crossing.point(line1, line2)
print(cr.pt)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
