% Generated by roxygen2 (4.0.2): do not edit by hand
\name{das.merge.acoustic}
\alias{das.merge.acoustic}
\title{Merge DAS Data Into Acoustic Detections}
\usage{
das.merge.acoustic(acoust.file, das.file, out.file)
}
\arguments{
\item{acoust.file}{filename of a .csv file of acoustic detections exported from Access database.}

\item{das.file}{filename of a DAS file.}

\item{out.file}{desired output .csv filename.}
}
\description{
Fill in sighting information for acoustic detections from DAS file.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

