% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das.read.R
\name{das.read}
\alias{das.read}
\title{Read DAS File}
\usage{
das.read(file)
}
\arguments{
\item{file}{filename of a DAS file.}
}
\description{
Reads a DAS file into a data.frame where each line is data for a specific
  event.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

