% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datum.R
\name{datum}
\alias{datum}
\title{Datum}
\usage{
datum(model = "wgs84")
}
\arguments{
\item{model}{character, specifying which model to use for ellipsoid model.
Options are: "WGS84", "GRS80", "Airy", "International", "Clarke", "GRS67",
or partial matches thereof (case-insensitive).}
}
\value{
vector of a, b, and f parameters.
}
\description{
Return parameters specifying ellipsoid datum model.
}
\note{
model parameters are based on distances in km.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

