% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Unbiased Estimate of Diversity}
\usage{
diversity(x)
}
\arguments{
\item{x}{character or numeric vector or factor}
}
\description{
Calculate unbiased estimate of diversity for a vector of items
}
\examples{
x <- sample(1:5, 100, replace = TRUE)
diversity(x)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
