\name{eiaMisc}
\alias{eiaMisc}
\docType{package}
\title{Set of miscellaneous functions}
\description{Collection of miscellaneous utility and analytical functions}
\details{
\tabular{ll}{
Package: \tab eiaMisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-08\cr
License: \tab GNU General Public License\cr
}
}
\author{Eric Archer<eric.archer@noaa.gov>}
\keyword{ package }