% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.method.p.R
\name{fisher.method.p}
\alias{fisher.method.p}
\title{Fisher's Method p-value}
\usage{
fisher.method.p(p.vals)
}
\arguments{
\item{p.vals}{vector of p-values.}
}
\description{
Calculate Fisher's method p-value to summarize a vector of 
  p-values based on a chi-squared distribution.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
