% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.p.R
\name{fisher.p}
\alias{fisher.p}
\title{Fisher's Method p-value}
\usage{
fisher.p(p)
}
\arguments{
\item{p}{vector of p-values.}
}
\description{
Calculate Fisher's method p-value to summarize a vector of 
  p-values based on a chi-squared distribution.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
