% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaParams.R
\name{gammaParams}
\alias{gammaParams}
\title{Calculate Gamma parameters}
\usage{
gammaParams(m, v)
}
\arguments{
\item{m}{mode}

\item{v}{variance}
}
\description{
Calculate Gamma rate and shape parameters from 
  mode and variance.
}
