% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.mean.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric Mean}
\usage{
geometric.mean(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{an optional numerical vector of weights the same length as \code{x}.}

\item{na.rm}{a logical value indicating whether NA values should be stripped 
before the computation proceeds.}
}
\description{
Calculates the geometric mean of a vector.
}
\examples{
x <- rlnorm(100)
mean(x)
median(x)
geometric.mean(x)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
