% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBiplot.R
\name{ggBiplot}
\alias{ggBiplot}
\title{ggBiplot}
\usage{
ggBiplot(
  pca,
  pcs = c(1, 2),
  mult.fac = 0.8,
  arrow.size = NULL,
  label.size = 3,
  plot = TRUE
)
}
\arguments{
\item{pca}{result from a call to \code{\link{princomp}} or \code{\link{prcomp}}.}

\item{pcs}{a two element vector giving the number or column names of 
the components to plot.}

\item{mult.fac}{multiplier factor for positions of points relative to extent of
plot (\code{0:1}).}

\item{arrow.size}{thickness of arrow lines. If \code{NULL} no arrows are shown.}

\item{label.size}{size of labels.}

\item{plot}{show the plot?}
}
\value{
the ggplot2 object is invisibly returned.
}
\description{
Create a biplot of a Principal Components Analysis using ggplot2.
}
\examples{
pca.princomp <- princomp(USArrests, cor = TRUE)
ggBiplot(pca.princomp)

pca.prcomp <- prcomp(USArrests, cor = TRUE)
ggBiplot(pca.prcomp)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
