% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBetween.R
\name{isBetween}
\alias{isBetween}
\title{Between}
\usage{
isBetween(x, a, b = NULL, include.ends = FALSE, na.convert = TRUE)
}
\arguments{
\item{x}{vector of numeric values to check.}

\item{a, b}{numeric values describing range.}

\item{include.ends}{logical. Should test include \code{a} and \code{b}? Is 
test > and < or >= and <= ?}

\item{na.convert}{logical. If \code{TRUE} and result of test is \code{NA} 
because either \code{x}, \code{a}, or \code{b} is \code{NA}, 
return \code{FALSE}, otherwise return \code{NA}.}
}
\description{
Is a numeric value between two other values?
}
\details{
Order of \code{a} and \code{b} does not matter. If \code{b} is 
  \code{NULL} the range will be taken from values in \code{a}.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
