% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lat.lon.list}
\alias{lat.lon.list}
\title{Get List of Latitude and Longitude Coordinates}
\usage{
lat.lon.list(area.poly)
}
\arguments{
\item{area.poly}{area of a polygon.}
}
\description{
Create a list of coordinates on the inputed polygon.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

