% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mail}
\alias{mail}
\title{Mail}
\usage{
mail(to, subject = "email from R", message = NULL, file = NULL)
}
\arguments{
\item{to}{recipient of email.}

\item{subject}{subject line of email.}

\item{message}{email content.}

\item{file}{file to be attached.}
}
\description{
Sends an email using the UNIX mail function.
}
\author{
Eric Archer \email{eric.archer@noaa.gov>}
}

