% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month2Season.R
\name{month2Season}
\alias{month2Season}
\title{Convert Months to Seasons}
\usage{
month2Season(x)
}
\arguments{
\item{x}{a vector of months from 1:12}
}
\description{
Convert numeric month to season: Winter = Dec-Feb, 
  Spring = Mar-May, Summer = Jun-Aug, Fall = Sep-Nov
}
\examples{
months <- sample(1:12, 10, rep = TRUE)
months
month2Season(months)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
