% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a Numeric Vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
a numeric vector of the same length as \code{x}.
}
\description{
Normalize a numeric vector to have a mean of zero and a standard deviation of one.\
}
\examples{
x <- runif(20, 50, 110)
x.norm <- normalize(x)
mean(x)
mean(x.norm)
sd(x)
sd(x.norm)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

