% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.arg.R
\name{one.arg}
\alias{one.arg}
\title{One Argument}
\usage{
one.arg(f)
}
\arguments{
\item{f}{a function.}
}
\description{
Does the function have just one argument?
}
\examples{
one.arg(mean)
one.arg(one.arg)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
