% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpPoint.R
\name{perpPoint}
\alias{perpPoint}
\title{Perpendicular Point}
\usage{
perpPoint(pts, line)
}
\arguments{
\item{pts}{two column matrix of points.}

\item{line}{two element vector giving intercept and slope of a line.}
}
\description{
Compute the perpendicular point between points and a line
  specified by an intercept and slope.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
