% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\docType{methods}
\name{round}
\alias{ceiling}
\alias{ceiling,data.frame-method}
\alias{ceiling.data.frame}
\alias{floor}
\alias{floor,data.frame-method}
\alias{floor.data.frame}
\alias{round}
\alias{round,data.frame-method}
\alias{round.data.frame}
\alias{siginif}
\alias{signif,data.frame-method}
\alias{signif.data.frame}
\alias{trunc}
\alias{trunc,data.frame-method}
\alias{trunc.data.frame}
\title{Rounding Numbers for Data Frames}
\usage{
\S4method{ceiling}{data.frame}(x)

\S4method{floor}{data.frame}(x)

\S4method{trunc}{data.frame}(x, ...)

\S4method{round}{data.frame}(x, digits = 0)

\S4method{signif}{data.frame}(x, digits = 6)
}
\arguments{
\item{x}{a data.frame with numeric columns.}

\item{...}{arguments to be passed to methods.}

\item{digits}{integer indicating the number of decimal places (\code{round})
or significant digits (\code{signif}) to be used. See \code{\link[base]{round}} for 
more details.}
}
\description{
Rounds numeric columns in data.frames
}
\details{
Takes a data.frame and returns a data.frame with the specified function 
  applied to each numeric column.
}
\examples{
data(mtcars)

round(mtcars, 0)

signif(mtcars, 2)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link[base]{round}}
}

