% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupClusters.R
\name{setupClusters}
\alias{setupClusters}
\title{Setup Clusters}
\usage{
setupClusters(num.cores = 1, max.cores = NULL)
}
\arguments{
\item{num.cores}{number of cores for multithreading. 
If \code{NULL}, the number used is set to the 
value of \code{parallel::detectCores() - 1}.}

\item{max.cores}{maximum number of cores to use.}
}
\value{
an object of class \code{c("SOCKcluster", "cluster")}.
}
\description{
Setup parallel clusters for different operating systems.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
