% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.symbols.R
\name{sex.symbols}
\alias{sex.symbols}
\title{Sex Symbols}
\usage{
sex.symbols(x, y, sex = 1, col = par("fg"), lwd = par("lwd"), cex = 1)
}
\arguments{
\item{x, y}{the \code{x} and \code{y} coordinates on the current plot.}

\item{sex}{a numeric vector containing the values 1 (male) or 2 (female). If of length one, then
value is recycled for all symbols.}

\item{col, lwd, cex}{color, line width, and character expansion for each point. \code{lwd} and 
\code{col} are recycled as necessary to cover all points. See \code{\link{par}} for more details.}
}
\description{
Plots male and female symbols on current plot.
}
\examples{
x <- runif(20, 0, 10)
y <- runif(20, 0, 200)
plot(x, y, type = "n")
sex.symbols(x, y, sex = 1:2, cex = 1.5, lwd = c(1.5, 4), col = c("blue", "red"))    

}
\author{
Tim Gerrodette \email{tim.gerrodette@noaa.gov}
}
