% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stouffer.p.R
\name{stouffer.p}
\alias{stouffer.p}
\title{Stouffer's Method p-value}
\usage{
stouffer.p(p, w = NULL)
}
\arguments{
\item{p}{vector of p-values.}

\item{w}{vector weights.}
}
\description{
Calculate Fisher's method p-value to summarize a vector of 
  p-values based on a chi-squared distribution.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
