% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfscMisc.R
\docType{package}
\name{swfscMisc}
\alias{swfscMisc}
\alias{swfscMisc-package}
\title{\code{swfscMisc} package}
\description{
SWFSC Miscellaneous Functions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EricArcher/swfscMisc}
  \item Report bugs at \url{https://github.com/EricArcher/swfscMisc/issues}
}

}
\author{
\strong{Maintainer}: Eric Archer \email{eric.ivan.archer@gmail.com}

}
