% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparent.R
\name{transparent}
\alias{transparent}
\title{Transparent Colors}
\usage{
transparent(col, percent = 50)
}
\arguments{
\item{col}{vector of colors as name, hexadecimal, or positive integer 
(see \code{\link{col2rgb}}).}

\item{percent}{percent of transparency (0 = solid, 100 = transparent).}
}
\description{
Return transparent form of a named color.
}
\examples{

pct <- seq(0, 100, by = 10)
plot(pct, pct, bg = transparent("red", pct), pch = 21, cex = 4, xlab = "X", ylab = "Y")

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
