% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.fisher.p.R
\name{weighted.fisher.p}
\alias{weighted.fisher.p}
\title{Weighted Fisher's Method p-value}
\usage{
weighted.fisher.p(p, w = NULL)
}
\arguments{
\item{p}{vector of p-values.}

\item{w}{vector weights.}
}
\description{
Calculate weighted Fisher's method p-value to summarize a vector of 
  p-values based on a chi-squared distribution.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
