% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.pad.R
\name{zero.pad}
\alias{zero.pad}
\title{Zero Pad Integers}
\usage{
zero.pad(x)
}
\arguments{
\item{x}{a vector of integers.}
}
\description{
Return character representation of integers that are zero-padded 
  to the left so all are the same length.
}
\examples{
x <- c(0, 1, 3, 4, 10) 
zero.pad(x)
x <- c(x, 11, 12, 100, 1000)
zero.pad(x)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
