% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert.distance}
\alias{convert.distance}
\title{Distance Conversion}
\usage{
convert.distance(x, from = "nm", to = "km")
}
\arguments{
\item{x}{numeric. The distance to be converted.}

\item{from,to}{character. Units to convert from and to. Can be "km" (kilometers),
"nm" (nautical miles), or "mi" (statute miles).}
}
\description{
Convert distances between kilometers, nautical miles, and statute miles.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

