% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das.spp.freq.R
\name{das.spp.freq}
\alias{das.spp.freq}
\title{Frequency of species from DAS file}
\usage{
das.spp.freq(x)
}
\arguments{
\item{x}{filename of a DAS file or a \code{data.frame} from \code{\link{das.read}}.}
}
\description{
Returns number of sightings of each species code in a DAS file.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
