% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datum.R
\name{datum}
\alias{datum}
\title{Datum}
\usage{
datum(model = c("wgs84", "grs80", "airy", "international", "clarke", "grs67"))
}
\arguments{
\item{model}{character, specifying which model to use for ellipsoid model. 
Options are: "wgs84", "grs80", "airy", "international", "clarke", "grs67", 
or partial matches thereof (case-sensitive).}
}
\value{
vector of a, b, and f parameters.
}
\description{
Return parameters specifying ellipsoid datum model.
}
\note{
Model parameters are based on distances in km.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
