% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.mean.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric Mean}
\usage{
geometric.mean(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Calculates the geometric mean of a vector.
}
\examples{
x <- rlnorm(100)
mean(x)
median(x)
geometric.mean(x)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
