% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBiplot.R
\name{ggBiplot}
\alias{ggBiplot}
\title{ggBiplot}
\usage{
ggBiplot(pca, x = 1, y = 2, mult.fac = 0.8, arrow.size = 1.5, label.size = 6)
}
\arguments{
\item{pca}{result from a call to \code{\link{princomp}}.}

\item{x, y}{the number or column names of the components to plot.}

\item{mult.fac}{multiplier factor for lengths of arrows from \code{0:1}.}

\item{arrow.size}{thickness of arrow lines.}

\item{label.size}{size of labels.}
}
\value{
the ggplot2 object is invisibly returned.
}
\description{
Plot a biplot of a Principal Components Analysis using ggplot2.
}
\examples{
pc.cr <- princomp(USArrests, cor = TRUE)
ggBiplot(pc.cr)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
