% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harmonic.mean.R
\name{harmonic.mean}
\alias{harmonic.mean}
\title{Harmonic Mean}
\usage{
harmonic.mean(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Harmonic Mean
}
\examples{
x <- rlnorm(100)
mean(x)
median(x)
harmonic.mean(x)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

