% Generated by roxygen2 (4.0.1): do not edit by hand
\name{one.arg}
\alias{one.arg}
\title{One Argument}
\usage{
one.arg(f)
}
\arguments{
\item{f}{a function.}
}
\description{
Does the function have just one argument?
}
\examples{
one.arg(mean)
one.arg(one.arg)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

