% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatterdens}
\alias{scatterdens}
\alias{scatterhist}
\title{Scatter Plot with Density Margins}
\usage{
scatterdens(x, y, dens.frac = 1/5, ...)

scatterhist(x, y, xlab = "", ylab = "", dens.frac = 1/5, ...)
}
\arguments{
\item{x,y}{vectors of points to plot.}

\item{dens.frac}{fraction of screen to be taken up by density plots on margins.}

\item{...}{Arguments to be passed to \link{plot}.}

\item{xlab,ylab}{labels for x and y axes.}
}
\description{
Scatter Plot with Density Margins
}
\examples{
x <- rnorm(100)
y <- rlnorm(100)
op <- par(ask = TRUE)
scatterdens(x, y, xlab = "x", ylab = "y")
par(op)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Original code by Ken Kleiman: \url{http://sas-and-r.blogspot.co.uk/2011/06/example-841-scatterplot-with-marginal.html}
}

