% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zero.pad}
\alias{zero.pad}
\title{Zero Pad Integers}
\usage{
zero.pad(x)
}
\arguments{
\item{x}{a vector of integers.}
}
\description{
Zero Pad Integers
}
\examples{
x <- c(0, 1, 3, 4, 10)
zero.pad(x)
x <- c(x, 11, 12, 100, 1000)
zero.pad(x)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

